faculty update

+++ app/Http/Controllers/Api/FacultyController.php
@@ -10,36 +10,16 @@ class FacultyController extends BaseController
 {
     public function listFaculty(){
         $results = Faculty::where('status',1)->get();
-
         return response()->json($results);
     }
 
     public function listFaculty_AllStatus(){
         $results = Faculty::get();
-
         return response()->json($results);
     }
 
     public function listFaculty_ById($facultyId){
         $results = Faculty::where('facultyId',$facultyId)->get();
-
-        return response()->json($results);
-    }
-
-    public function Search_Faculty(Request $request){
-        $results = Faculty::where('facultyId','like','%'.$request->facultyId.'%')
-                            ->Where('facultyName','like','%'.$request->facultyName.'%')
-                            ->get();
-
-        return response()->json($results);
-    }
-
-    public function Search_Faculty_Active(Request $request){
-        $results = Faculty::where('facultyId','like','%'.$request->facultyId.'%')
-                            ->Where('facultyName','like','%'.$request->facultyName.'%')
-                            ->where('status','1')
-                            ->get();
-
         return response()->json($results);
     }
 
@@ -49,7 +29,6 @@ class FacultyController extends BaseController
         $newFaculty->facultyName   = $request->facultyName;
         $newFaculty->status         = 1;
         $newFaculty->save();
-
         return response()->json($request);
     }
 
@@ -58,38 +37,39 @@ class FacultyController extends BaseController
             ->update([
                 'facultyName' => $request->facultyName,
             ]);
-
         $resp = array('status'=>1, 'message'=>'Edit success');
         return response()->json($resp);
     }
 
-    public function changstatusFaculty(Request $request){
-        $results = Department::where('facultyId',$request->facultyId)->get();
-        if($results == 0){
-
-        }
-        Faculty::where('facultyId',$request->facultyId)
-            ->update([
-                'status' => $request->status
-            ]);
-
-        $resp = array('status'=>1, 'message'=>'Change Status Success');
-        return response()->json($request);
+    public function changstatusFaculty(Request $request){ 
+        $result_subject = Faculty::join('subjects','subjects.facultyId','=','faculty.facultyId')
+                            ->where('faculty.facultyId',$request->facultyId)->get();
+        $result_group = Faculty::join('groups','groups.facultyId','=','faculty.facultyId')
+                            ->where('faculty.facultyId',$request->facultyId)->get();
+        $resp = array('status'=>1, 'message'=>'');
+            if($result_subject == '[]' && $result_group == '[]'){
+                Faculty::where('facultyId',$request->facultyId)
+                        ->update(['status' => $request->status]);
+                $resp = array('status'=>1, 'message'=>'เปลี่ยนสถานะคณะสำเร็จ');
+            }else{
+                $resp = array('status'=>0, 'message'=>'เปลี่ยนสถานะคณะไม่สำเร็จ');
+            }
+        return response()->json($resp);
     }
 
     public function deleteFaculty(Request $request){
-        $deleteFaculty = Faculty::where('facultyId',$request->facultyId)->delete();
-
-        return response()->json($request);
-    }
-
-    public function checkDelFaculty(Request $request){
-        $results = Department::where('facultyId',$request->facultyId)->get();
-        if($results != '[]'){
-            $resp = array('status'=>0, 'message'=>'ไม่สามารถทำรายการลบข้อมูลคณะรหัส : '.$request->facultyId.' ได้');
+        $result_subject = Faculty::join('subjects','subjects.facultyId','=','faculty.facultyId')
+                            ->where('faculty.facultyId',$request->facultyId)->get();
+        $result_group = Faculty::join('groups','groups.facultyId','=','faculty.facultyId')
+                            ->where('faculty.facultyId',$request->facultyId)->get();
+        $resp = array('status'=>1, 'message'=>'');
+        if($result_subject == '[]' && $result_group == '[]'){
+            $deleteFaculty = Faculty::where('facultyId',$request->facultyId)->delete();
+            $resp = array('status'=>1, 'message'=>'ลบข้อมูลคณะสำเร็จ');
         }else{
-            $resp = array('status'=>1, 'message'=>'ยืนยันการลบข้อมูลคณะรหัส : '.$request->facultyId.' หรือไม่');
+            $resp = array('status'=>0, 'message'=>'ลบข้อมูลคณะไม่สำเร็จ');
         }
         return response()->json($resp);
     }
+
 }
